+++ apps/web/app/tamagui-example/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { TamaguiButton } from '@repo/ui/tamagui-button'
-import { TamaguiCard } from '@repo/ui/tamagui-card'
+import { TamaguiButton } from '@repo/ui/button'
+import { TamaguiCard } from '@repo/ui/card'
 import { useState } from 'react'
 
 export default function TamaguiExample() {
