+++ packages/ui/src/button.tsx
@@ -1,21 +1,49 @@
-'use client'
+import type { GetProps } from 'tamagui'
+import { styled, Button as TamaguiButtonBase } from 'tamagui'
 
-import type { ReactNode } from 'react'
+export const StyledButton = styled(TamaguiButtonBase, {
+  name: 'StyledButton',
+  variants: {
+    variant: {
+      primary: {
+        bg: '$background',
+        color: '$color',
+      },
+      secondary: {
+        bg: '$backgroundPress',
+        color: '$color',
+      },
+      ghost: {
+        bg: 'transparent',
+        borderWidth: 1,
+        borderColor: '$borderColor',
+        color: '$color',
+      },
+    },
+    customSize: {
+      small: {
+        px: '$2',
+        py: '$1',
+        fontSize: '$1',
+      },
+      medium: {
+        px: '$3',
+        py: '$2',
+        fontSize: '$3',
+      },
+      large: {
+        px: '$4',
+        py: '$3',
+        fontSize: '$5',
+      },
+    },
+  } as const,
+  defaultVariants: {
+    variant: 'primary',
+    customSize: 'medium',
+  },
+})
 
-interface ButtonProps {
-  children: ReactNode
-  className?: string
-  appName: string
-}
+export type StyledButtonProps = GetProps<typeof StyledButton>
 
-export const Button = ({ children, className, appName }: ButtonProps) => {
-  return (
-    <button
-      type='button'
-      className={className}
-      onClick={() => alert(`Hello from your ${appName} app!`)}
-    >
-      {children}
-    </button>
-  )
-}
+export { StyledButton as TamaguiButton }
