+++ packages/ui/src/card.tsx
@@ -1,27 +1,49 @@
-import type { JSX } from 'react'
+import type { ReactNode } from 'react'
+import type { CardProps } from 'tamagui'
+import { Card, H2, Paragraph, XStack, YStack } from 'tamagui'
 
-export function Card({
-  className,
-  title,
-  children,
-  href,
-}: {
-  className?: string
-  title: string
-  children: React.ReactNode
-  href: string
-}): JSX.Element {
+interface TamaguiCardProps extends CardProps {
+  title?: string
+  description?: string
+  footer?: ReactNode
+}
+
+export function TamaguiCard({ title, description, footer, children, ...props }: TamaguiCardProps) {
   return (
-    <a
-      className={className}
-      href={`${href}?utm_source=create-turbo&utm_medium=basic&utm_campaign=create-turbo"`}
-      rel='noopener noreferrer'
-      target='_blank'
+    <Card
+      elevate
+      bordered
+      animation='bouncy'
+      size='$4'
+      padding='$4'
+      gap='$3'
+      hoverStyle={{
+        scale: 0.98,
+        borderColor: '$borderColor',
+      }}
+      pressStyle={{
+        scale: 0.97,
+      }}
+      {...props}
     >
-      <h2>
-        {title} <span>-&gt;</span>
-      </h2>
-      <p>{children}</p>
-    </a>
+      <YStack gap='$2'>
+        {title && (
+          <H2 size='$6' color='$color'>
+            {title}
+          </H2>
+        )}
+        {description && (
+          <Paragraph size='$3' color='$colorPress'>
+            {description}
+          </Paragraph>
+        )}
+        {children}
+      </YStack>
+      {footer && (
+        <XStack pt='$3' borderTopWidth={1} borderTopColor='$borderColor'>
+          {footer}
+        </XStack>
+      )}
+    </Card>
   )
 }
