+++ packages/ui/src/index.ts
@@ -1,3 +1,3 @@
+export * from './button'
+export * from './card'
 export { default as tamaguiConfig } from './tamagui.config'
-export * from './tamagui-button'
-export * from './tamagui-card'
